examples/openai-agents-trace-evidence/requirements.txt
openai-agents==0.13.5
